[Modify] Entity 데이터 모델로 변환
